Move PagurePullRequestCommentCoprBuildHandler to new pagure_handlers.py
